requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pygame==2.5.2
PyOpenGL==3.1.7
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.11.3
six==1.16.0
zipp==3.17.0
